Split other-window commands to a separate file
